feat(client): add social media icons and new sections for projects and call to action
