+++ apps/client/src/icons/LinkedInIcon.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const LinkedInIcon: React.FC = () => (
+  <svg width="20" height="20" fill="currentColor" viewBox="0 0 24 24">
+    <path d="M16 8a6 6 0 016 6v7h-4v-7a2 2 0 00-2-2 2 2 0 00-2 2v7h-4v-7a6 6 0 016-6zM2 9h4v12H2z"></path>
+    <circle cx="4" cy="4" r="2"></circle>
+  </svg>
+);
+
+export default LinkedInIcon;
