+++ apps/client/src/icons/TwitterIcon.tsx
@@ -1,0 +1,94 @@
+import React from 'react';
+
+const TwitterIcon: React.FC = () => (
+  <svg width="20" height="20" fill="currentColor" viewBox="0 0 24 24">
+    <path d="M23 3a10 10 0 01-3 1 5 5 0 00-8.5 4.5v1A10 10 0 013 4s-4 9 5 13a11 11 0 01-7 2c9 5 20 0 20-11.5a6 6 0 00-.1-1c1-.7 1.8-1.8 2.1-3z"></path>
+  </svg>
+);
+
+export default TwitterIcon;
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
